pyproject.toml
[project]
name = "bb-review"
version = "0.1.0"
description = "AI-powered code review system for Review Board"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "BB Review Team"}
]
keywords = ["code-review", "review-board", "ai", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "pydantic>=2.0",
    "GitPython>=3.1.0",
    "requests>=2.28.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[project.scripts]
bb-review = "bb_review.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bb_review"]

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"bb_review/cli/__init__.py" = ["E402", "F401"]  # Intentional late imports for CLI subcommands

[tool.ruff.lint.flake8-quotes]
avoid-escape = false
inline-quotes = "single"
docstring-quotes = "single"
multiline-quotes = "single"

[tool.ruff.lint.mccabe]
max-complexity = 16

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = true
lines-after-imports = 2

[tool.ruff.lint.pylint]
max-args = 8
max-branches = 30
max-statements = 100
max-returns = 7

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]
